+++ src/LocalProxy.App/AppShell.xaml.cs
@@ -1,0 +1,9 @@
+﻿namespace LocalProxy.App;
+
+public partial class AppShell : Shell
+{
+	public AppShell()
+	{
+		InitializeComponent();
+	}
+}